Show classes found

+++ ThePersonalTrainerClub/app/src/main/java/es/exceptioncoders/thepersonaltrainerclub/model/model/LocationModel.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class LocationModel(
         val id: String?,
         val type: String,
-        val coordinates: Array<Float>,
+        val coordinates: Array<Double>,
         val description: String?
 ): Serializable
